+++ src/compiler/types/mod.rs
@@ -1,0 +1,6 @@
+mod lines_0_4000;
+pub use lines_0_4000::*;
+mod lines_4000_8000;
+pub use lines_4000_8000::*;
+mod lines_8000_end;
+pub use lines_8000_end::*;
